Moved performance measurment and ESP specific definitions to umm_performance.h/cpp. Removed testing asserts.
