+++ cores/esp8266/umm_malloc/umm_malloc.cpp
@@ -495,32 +495,12 @@
 #include <string.h>
 #include <pgmspace.h>
 
+#define _UMM_MALLOC_CPP
 #include "umm_malloc.h"
 
 #include "umm_malloc_cfg.h"   /* user-dependent */
-#if defined(DEBUG_ESP_PORT)
-#include <assert.h>
-#endif
 
 extern "C" {
-
-#ifdef UMM_CRITICAL_PERIOD_ANALYZE
-static struct _UMM_TIME_STATS time_stats = {
-  {0xFFFFFFFF, 0U, 0U, 0U},
-  {0xFFFFFFFF, 0U, 0U, 0U},
-  {0xFFFFFFFF, 0U, 0U, 0U},
-  {0xFFFFFFFF, 0U, 0U, 0U} };
-
-bool ICACHE_FLASH_ATTR get_umm_get_perf_data(struct _UMM_TIME_STATS *p, size_t size) {
-    if (p && size != 0) {
-        uint32_t save_ps = xt_rsil(DEFAULT_CRITICAL_SECTION_INTLEVEL);
-        memcpy(p, &time_stats, size);
-        xt_wsr_ps(save_ps);
-        return true;
-    }
-    return false;
-}
-#endif
 
 // From UMM, the last caller of a malloc/realloc/calloc which failed:
 extern void *umm_last_fail_alloc_addr;
@@ -575,7 +555,7 @@ Changes for July 2019:
   test case that shows an issue yet; however, I and others think they are or
   have had crashes related to this.
 
-  I have added code that adjusts the running free heap number from _umm_malloc,
+  I have added code that updates a current free heap value from _umm_malloc,
   _umm_realloc, and _umm_free. Removing the need to do a long interrupts
   disabled calculation via umm_info.
 
@@ -584,58 +564,6 @@ Changes for July 2019:
   called with INTLEVEL set.
 
  */
-
-#if defined(DEBUG_ESP_PORT) || defined(DEBUG_ESP_ISR)
-#define DEBUG_IRAM_ATTR ICACHE_RAM_ATTR
-/*
-  Printing from the malloc routines is tricky. Since a lot of library calls
-  will want to do malloc.
-
-  Objective:  To be able to print "last gasp" diagnostic messages
-  when interrupts are disabled and w/o availability of heap resources.
-*/
-
-// ROM _putc1, ignores CRs and sends CR/LF for LF, newline.
-// Always returns character sent.
-int constexpr (*_rom_putc1)(int) = (int (*)(int))0x40001dcc;
-void uart_buff_switch(uint8_t);
-
-int DEBUG_IRAM_ATTR _isr_safe_printf_P(const char *fmt, ...) __attribute__((format(printf, 1, 2)));
-int DEBUG_IRAM_ATTR _isr_safe_printf_P(const char *fmt, ...) {
-#ifdef DEBUG_ESP_PORT
-#define VALUE(x) __STRINGIFY(x)
-  // Preprocessor and compiler together will optimize away the if.
-  if (strcmp("Serial1", VALUE(DEBUG_ESP_PORT)) == 0) {
-    uart_buff_switch(1U);
-  } else {
-    uart_buff_switch(0U);
-  }
-#else
-  uart_buff_switch(0U); // Side effect, clears RX FIFO
-#endif
-  /*
-    To use ets_strlen() and ets_memcpy() safely with PROGMEM, flash storage,
-    the PROGMEM address must be word (4 bytes) aligned. The destination
-    address for ets_memcpy must also be word-aligned. We also round the
-    buf_len up to the nearest word boundary. So that all transfers will be
-    whole words.
-  */
-  size_t str_len = ets_strlen(fmt);
-  size_t buf_len = (str_len + 1 + 3) & ~0x03U;
-  char ram_buf[buf_len] __attribute__ ((aligned(4)));
-  ets_memcpy(ram_buf, fmt, buf_len);
-  va_list argPtr;
-  va_start(argPtr, fmt);
-  int result = ets_vprintf(_rom_putc1, ram_buf, argPtr);
-  va_end(argPtr);
-  return result;
-}
-
-#define printf(fmt, ...) _isr_safe_printf_P(PSTR(fmt), ##__VA_ARGS__)
-#else
-// Macro to place constant strings into PROGMEM and print them properly
-#define printf(fmt, ...) printf(PSTR(fmt), ## __VA_ARGS__ )
-#endif
 
 /* -- dbglog {{{ */
 
@@ -1205,10 +1133,6 @@ void ICACHE_FLASH_ATTR *umm_info( void *ptr, int force ) {
       ummHeapInfo.maxFreeContiguousBlocks = curBlocks;
     }
   }
-#if defined(DEBUG_ESP_PORT)
-  // This is a good place to verify we are calculating correctly
-  assert(ummHeapFreeBlocks == ummHeapInfo.freeBlocks);
-#endif
 
   DBG_LOG_FORCE( force, "|0x%08lx|B %5d|NB %5d|PB %5d|Z %5d|NF %5d|PF %5d|\n",
       (unsigned long)(&UMM_BLOCK(blockNo)),
